+++ CSK_Module_MultiIOLinkSMI/scripts/Communication/MultiIOLinkSMI/helper/funcs.lua
@@ -196,9 +196,6 @@ local function checkParameters(content, defaultTable)
     elseif content[key] == nil then
       _G.logger:info(nameOfModule .. ": Missing parameter '" .. tostring(key) .. "'. Adding default value '" .. tostring(defaultTable[key]) .. "'")
       content[key] = defaultTable[key]
-      if key == 'cameraNo' then
-        _G.logger:warning(nameOfModule .. ": '" .. tostring(key) .. "' is a major parameter! Default value might not work and needs to be edited!")
-      end
     end
   end
   return content